pyproject.toml
[tool.poetry]
name = "repl_nix_saleinvoicewebsite"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.2"
mysql-connector-python = "^8.1.0"
numpy = "^1.25.1"
pandas = "^2.0.3"
SQLAlchemy = "^2.0.19"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
